requirements.txt
torch~=1.13.1
torchvision~=0.14.1
matplotlib~=3.6.3
Pillow~=9.0.1
numpy~=1.23.0